requirements.txt
# requirements.txt

# install with pip install -r requirements.txt

# Standard library modules, no need to include in requirements.txt
# threading
# subprocess
# platform
# re
# math
# os



## backend dependencies

# for downloading tiles
requests

# PIL for image processing
Pillow

# png to pdf
img2pdf


## frontend dependencies

# flask
flask
flask-socketio

# pywebview
pywebview

# screeninfo
screeninfo



## Nuitka
nuitka